remove v2 of 3b lesson script
